requirements.txt
requests==2.30.0
flake8==6.0.0
python-telegram-bot==20.3
python-dotenv==1.0.0
psycopg2-binary==2.9.6
